Add client-side searching, sorting and limiting for incompatible datasources to new DataFetch model
