+++ packages/client/src/utils/fetch/ViewFetch.js
@@ -1,16 +1,7 @@
-import TableFetch from "./TableFetch.js"
+import DataFetch from "./DataFetch.js"
 import { fetchTableDefinition, fetchViewData } from "api"
 
-export default class ViewFetch extends TableFetch {
-  SupportsSearch = false
-  SupportsSort = false
-  SupportsPagination = false
-
-  /**
-   * Fetches the schema for a view
-   * @param datasource the view datasource config
-   * @return {object} the view schema
-   */
+export default class ViewFetch extends DataFetch {
   static async getSchema(datasource) {
     if (!datasource?.tableId) {
       return null
@@ -19,9 +10,6 @@ export default class ViewFetch extends TableFetch {
     return this.enrichSchema(table?.views?.[datasource.name]?.schema)
   }
 
-  /**
-   * Fetches a single page of data from the remote resource
-   */
   async getData() {
     const { datasource } = this.options
     const res = await fetchViewData(datasource)
@@ -30,5 +18,3 @@ export default class ViewFetch extends TableFetch {
     }
   }
 }
-
-ViewFetch.getSchema()
